Adjust how start/stop autocmds are configured
Before this change, the christmas_glitter_tick() timer function would
continue looping even if the christmas-tree buffer was not visible but
still loaded. This is not a huge problem since it only runs once per
second, but is still not ideal. This change configures the timer to stop
when the buffer is no longer visible (BufWinLeave) and to resume when
the buffer becomes visible (BufWinEnter).

This also refactors the autocmds into a named augroup. This is probably
not a big issue either, but it makes it easy to clear the group so we
are safe if we somehow execute ftplugin twice for the buffer.

Test: Seems to work on my machine (nvim 0.6). Even if I load the same
buffer into multiple splits, it seems to only have 1 timer going (so all
the splits are synchronized, changing colors once per second as
expected).
Test: move the buffer to the background and call `:hi
christmastreeGlitterLarge` multiple times (over the course of several
seconds), verify the highlight linking remains the same